Give the dashboard's attention queue its own module and its own verbs
managerNotifications sat in membership.functions.ts because membership
windows were the only thing that could need a manager's attention. It is
about to have a second source, so move it to
manager-notifications.functions.ts, whose job is composing sources and
deciding what order a manager meets them in. Membership keeps the rule it
contributes; one import site moved.

The button label was hardcoded as "Fix it" for every row, which is right
for unset training dates and wrong for anything that is not a fault. Make
actionLabel a required field on ManagerNotification so the compiler points
at every construction site rather than letting a new kind of notification
inherit a verb that does not fit it. The existing builder passes "Fix it",
so nothing on screen changes.
